Ensure the GetAssemblyTest runs at the end by making it sequencial
