+++ tracer/test/Datadog.Trace.DuckTyping.Tests/GetAssemblyTestsCollection.cs
@@ -1,0 +1,14 @@
+// <copyright file="GetAssemblyTestsCollection.cs" company="Datadog">
+// Unless explicitly stated otherwise all files in this repository are licensed under the Apache 2 License.
+// This product includes software developed at Datadog (https://www.datadoghq.com/). Copyright 2017 Datadog, Inc.
+// </copyright>
+
+using Xunit;
+
+namespace Datadog.Trace.DuckTyping.Tests
+{
+    [CollectionDefinition(nameof(GetAssemblyTestsCollection), DisableParallelization = true)]
+    public class GetAssemblyTestsCollection
+    {
+    }
+}
